+++ apps/i2ptunnel/java/src/net/i2p/i2ptunnel/access/AccessFilter.java
@@ -232,7 +232,7 @@ class AccessFilter implements StatefulConnectionFilter {
     }
 
     private void purge() {
-        long olderThan = context.clock().now() - definition.getPurgeSeconds() * 1000;
+        long olderThan = context.clock().now() - (long) definition.getPurgeSeconds() * 1000;
 
         synchronized(knownDests) {
             for (DestTracker tracker : knownDests.values()) {
