+++ router/java/src/net/i2p/router/HashPatternDetector.java
@@ -409,10 +409,10 @@ public class HashPatternDetector implements Serializable {
         try {
             if (prop.length() > 1 && (prop.endsWith("m") || prop.endsWith("M"))) {
                 int minutes = Integer.parseInt(prop.substring(0, prop.length() - 1));
-                return minutes * 60 * 1000L;
+                return (long) minutes * 60 * 1000L;
             } else {
                 int hours = Integer.parseInt(prop);
-                return hours * 60 * 60 * 1000L;
+                return (long) hours * 60 * 60 * 1000L;
             }
         } catch (NumberFormatException e) {
             if (_log.shouldWarn())
